requirements.txt
streamlit
pandas
Pillow
python-docx